requirements.txt
websocket-client>=1.6.0
websockets>=11.0
pandas>=2.0.0
numpy>=1.24.0
statsmodels>=0.14.0
scipy>=1.10.0
scikit-learn>=1.3.0
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
anthropic>=0.18.0
plotly>=5.0.0
kaleido>=0.2.1
httpx>=0.24.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
httpx>=0.24.0
fastapi[all]>=0.100.0
uvicorn>=0.23.0
websocket-client>=1.6.0
faker>=20.0.0
sentry-sdk>=2.0.0
